\name{read.weka}
\alias{read.weka}
\title{
Read WEKA ARFF files
}
\description{
Import WEKA ARFF files and convert coerce data to data.frame
}
\usage{
read.weka(filename, asList = FALSE)
}
\arguments{
  \item{filename}{
name of the ARFF file
}
  \item{asList}{
logical. If TRUE, a list including the dataframe and the WEKA relation is returned}
}
\examples{
#df <- read.weka("kkkk")
#head(df)
}
